refactor: rename invite files to mail-history
