+++ src/auth/auth.module.ts
@@ -14,7 +14,7 @@ import { IsNotExist } from 'src/utils/validators/is-not-exists.validator';
 import { CoreBankService } from 'src/core-bank/core-bank.service';
 import { IsValidBankCodeConstraint } from 'src/banks/validators/is-valid-bank-code.validator';
 import { CoreBankModule } from 'src/core-bank/core-bank.module';
-import { InviteModule } from 'src/invite/invite.module';
+import { MailHistoryModule } from 'src/mail-history/mail-history.module';
 
 @Module({
   imports: [
@@ -23,7 +23,7 @@ import { InviteModule } from 'src/invite/invite.module';
     PassportModule,
     MailModule,
     CoreBankModule,
-    InviteModule,
+    MailHistoryModule,
     JwtModule.registerAsync({
       imports: [ConfigModule],
       inject: [ConfigService],
