+++ src/auth/auth.service.ts
@@ -7,9 +7,9 @@ import * as crypto from 'crypto';
 import { CoreBankService } from 'src/core-bank/core-bank.service';
 import { UpdateCoreBankInterface } from 'src/core-bank/interfaces/update-core-bank.interface';
 import { ForgotService } from 'src/forgot/forgot.service';
-import { InviteStatusEnum } from 'src/invite-statuses/invite-status.enum';
-import { MailHistory } from 'src/invite/entities/invite.entity';
-import { MailHistoryService } from 'src/invite/mail-history.service';
+import { InviteStatusEnum } from 'src/mail-history-statuses/mail-history-status.enum';
+import { MailHistory } from 'src/mail-history/entities/mail-history.entity';
+import { MailHistoryService } from 'src/mail-history/mail-history.service';
 import { MailData } from 'src/mail/interfaces/mail-data.interface';
 import { MailService } from 'src/mail/mail.service';
 import { Role } from 'src/roles/entities/role.entity';
