+++ src/cron-jobs/cron-jobs.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CronJobsService } from './cron-jobs.service';
 import { ScheduleModule } from '@nestjs/schedule';
-import { InviteModule } from 'src/invite/invite.module';
+import { MailHistoryModule } from 'src/mail-history/mail-history.module';
 import { MailModule } from 'src/mail/mail.module';
 import { SettingsModule } from 'src/settings/settings.module';
 import { ConfigModule } from '@nestjs/config';
@@ -15,7 +15,7 @@ import { MailCountModule } from 'src/mail-count/mail-count.module';
     ScheduleModule.forRoot(),
     ConfigModule,
     SettingsModule,
-    InviteModule,
+    MailHistoryModule,
     MailModule,
     UsersModule,
     JaeModule,
