+++ src/database/seeds/mail-history-status/mail-history-status-seed.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { InviteStatusSeedService } from './invite-status-seed.service';
-import { InviteStatus } from 'src/invite-statuses/entities/invite-status.entity';
+import { InviteStatusSeedService } from './mail-history-status-seed.service';
+import { InviteStatus } from 'src/mail-history-statuses/entities/mail-history-status.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([InviteStatus])],
