+++ src/mail-history-statuses/entities/mail-history-status.entity.ts
@@ -1,7 +1,7 @@
 import { Column, Entity, PrimaryColumn } from 'typeorm';
 import { ApiProperty } from '@nestjs/swagger';
 import { EntityHelper } from 'src/utils/entity-helper';
-import { InviteStatusEnum } from '../invite-status.enum';
+import { InviteStatusEnum } from '../mail-history-status.enum';
 import { Enum } from 'src/utils/enum';
 
 @Entity()
