+++ src/mail-history-statuses/interfaces/invite-status.interface.ts
