+++ src/mail-history/pipes/mail-history-hash-exists.pipe.ts
