+++ src/users/users.module.ts
@@ -5,10 +5,10 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { IsExist } from 'src/utils/validators/is-exists.validator';
 import { IsNotExist } from 'src/utils/validators/is-not-exists.validator';
-import { InviteModule } from 'src/invite/invite.module';
+import { MailHistoryModule } from 'src/mail-history/mail-history.module';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([User]), InviteModule],
+  imports: [TypeOrmModule.forFeature([User]), MailHistoryModule],
   controllers: [UsersController],
   providers: [IsExist, IsNotExist, UsersService],
   exports: [UsersService],
