+++ src/users/users.service.ts
@@ -3,10 +3,10 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { plainToClass } from 'class-transformer';
 import { validate } from 'class-validator';
 import { Request } from 'express';
-import { InviteStatus } from 'src/invite-statuses/entities/invite-status.entity';
-import { InviteStatusEnum } from 'src/invite-statuses/invite-status.enum';
-import { MailHistory } from 'src/invite/entities/invite.entity';
-import { MailHistoryService } from 'src/invite/mail-history.service';
+import { InviteStatus } from 'src/mail-history-statuses/entities/mail-history-status.entity';
+import { InviteStatusEnum } from 'src/mail-history-statuses/mail-history-status.enum';
+import { MailHistory } from 'src/mail-history/entities/mail-history.entity';
+import { MailHistoryService } from 'src/mail-history/mail-history.service';
 import { Role } from 'src/roles/entities/role.entity';
 import { RoleEnum } from 'src/roles/roles.enum';
 import { Status } from 'src/statuses/entities/status.entity';
